index<>single 페이지로 분리 및 index 페이지 디자인
